project/service/webscrape/requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
Flask-Cors==4.0.0
requests 
beautifulsoup4
googlesearch-python